Relations Map

+++ Voyage-Service/src/main/java/pfe/emsi/covoiturage/voyageservice/Controllers/VoyageDTO.java
@@ -4,6 +4,6 @@ import pfe.emsi.covoiturage.voyageservice.Entities.Coordinates;
 
 import java.time.LocalDateTime;
 
-public record VoyageDTO(float destinationX, float destinationY, float emplacementX, float emplacementY,String labelDestination,String labelEmplacement, float Tarif,
+public record VoyageDTO(double destinationX, double destinationY, double emplacementX, double emplacementY,String labelDestination,String labelEmplacement, float Tarif,
                         int Duree, int Distance, String driver, LocalDateTime time) {
 }
